image-generator-app/backend/requirements.txt
Flask==2.0.2
requests==2.26.0
celery==5.2.1
redis==4.0.2
minio==7.1.0
Werkzeug==2.2.2
flask-cors==4.0.1
flask[async]
python-dotenv==1.0.1